pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "discoverykastle-server"
version = "0.1.0"
description = "Autonomous network discovery, security assessment, and documentation platform."
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",

    # Validation / settings
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # HTTP client (notifications, NetBox, webhooks)
    "httpx>=0.27",

    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",

    # Crypto (CA / mTLS)
    "cryptography>=42",

    # Redis
    "redis>=5.0",

    # Templating (doc builder)
    "jinja2>=3.1",

    # Optional: YAML config for external modules
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

# pip install 'discoverykastle-server[graylog]'
# Required when DKASTLE_GRAYLOG_HOST is set.
graylog = [
    "graypy>=2.1",
]

# pip install 'discoverykastle-server[ai]'
# Required when DKASTLE_AI_ENABLED=true.
# Only install if you explicitly need AI-powered contextual CVE triage.
ai = [
    "anthropic>=0.40",
]

# pip install 'discoverykastle-server[webpush]'
# Required when DKASTLE_WEBPUSH_ENABLED=true.
webpush = [
    "pywebpush>=2.0",
]

# ------------------------------------------------------------------
# Entry point group for external modules
# External packages declare their module class here:
#
#   [project.entry-points."discoverykastle.modules"]
#   my-module = "my_package.module:Module"
# ------------------------------------------------------------------
[project.entry-points."discoverykastle.modules"]
# (built-ins are loaded directly by the loader, not via entry points)

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "modules*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
